New features        : Problems solved     : - Fixed I32/I64 incompatibility Bugzilla-Bug-ID     : Additional comments :
git-svn-id: svn://doncamillo.mathematik.tu-dortmund.de/Featflow2@4547 4b12f95d-9313-0410-afc9-bd26427041f1

+++ applications/anisotropicdiffusion/src/anisotropicdiffusion_method1.f90
@@ -132,7 +132,7 @@ contains
     real(DP) :: dmeshDistortion
     
     ! Element type of the discretisation
-    integer :: ieltype
+    integer(I32) :: ieltype
     
     ! Whether to convert to triangle mesh
     integer :: iconvertToTriangleMesh
